Migrate to new unit test standard
