+++ src/Flagscript/Flagscript.csproj
@@ -22,7 +22,7 @@
     <Description>Foundational Classes for the Flagscript Framework</Description>
     <AssemblyVersion>3.1.0</AssemblyVersion>
     <FileVersion>3.1.0</FileVersion>
-    <InformationalVersion>3.1.0-beta</InformationalVersion>
+    <InformationalVersion>3.1.0</InformationalVersion>
     <PackageId>Flagscript</PackageId>
     <Product>Flagscript</Product>
   </PropertyGroup>
