--- test/Flagscript.Test/Flagscript.Test.csproj
@@ -1,36 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <ReleaseVersion>3.0.1</ReleaseVersion>
-    <PackageVersion>3.0.1</PackageVersion>
-    <Authors>Greg Kaestle, Flagscript</Authors>
-    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <Summary>Greg Kaestle, Flagscript</Summary>
-    <Title>Flagscript.Test</Title>
-    <Description>Unit Tests for Flagscript.</Description>
-    <PackageIconUrl>https://raw.githubusercontent.com/flagscript/Flagscript/master/flagscript-nuget.png</PackageIconUrl>
-    <NeutralLanguage>en-US</NeutralLanguage>
-    <PackageLicenseUrl>https://github.com/flagscript/Flagscript/blob/master/LICENSE.md</PackageLicenseUrl>
-    <PackageProjectUrl>https://github.com/flagscript/Flagscript</PackageProjectUrl>
-    <PackageTags>flagscript, dotnet, dotnet-core, dotnet-standard, dotnet-library</PackageTags>
-    <Owners>Flagscript</Owners>
-    <PackageId>Flagscript.Test</PackageId>
-    <PackageReleaseNotes>Documentation Fixes.</PackageReleaseNotes>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
-    <PackageReference Include="xunit" Version="2.4.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
-    <PackageReference Include="NuGet.Build.Packaging" Version="0.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Flagscript\Flagscript.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Extensions\" />
-    <Folder Include="Extensions\Dotnet\" />
-  </ItemGroup>
-</Project>
