+++ test/Flagscript.Unit.Tests/Extensions/Dotnet/TestStruct.cs
@@ -1,0 +1,11 @@
+﻿namespace Flagscript.Unit.Tests.Extensions.Dotnet
+{
+
+	/// <summary>
+	/// Test struct for <see cref="EnumExtensionsTest.TestGetEnumDescriptionNonEnumType"/>.
+	/// </summary>
+	public struct TestStruct
+	{
+	}
+
+}
