+++ test/Flagscript.Unit.Tests/Flagscript.Unit.Tests.csproj
@@ -1,0 +1,25 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <IsPackable>false</IsPackable>
+    <Authors>Greg Kaestle, Flagscript</Authors>
+    <Company>Flagscript</Company>
+    <Copyright>©2019 Flagscript</Copyright>
+    <RepositoryUrl>https://github.com/flagscript/Flagscript</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <Owners>Greg Kaestle, Flagscript</Owners>
+    <Summary>Unit tests for Flagscript</Summary>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
+    <PackageReference Include="xunit" Version="2.4.0" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Flagscript\Flagscript.csproj" />
+  </ItemGroup>
+ 
+</Project>
