+++ test/Flagscript.Unit.Tests/FlagscriptConfigurationExceptionTest.cs
@@ -1,0 +1,26 @@
+﻿using Xunit;
+
+namespace Flagscript.Unit.Tests
+{
+
+	/// <summary>
+	/// Unit tests for <see cref="FlagscriptConfigurationException"/>.
+	/// </summary>
+	public class FlagscriptConfigurationExceptionTest
+	{
+
+		/// <summary>
+		/// Tests that <see cref="FlagscriptConfigurationException"/> is of type 
+		/// <see cref="FlagscriptException"/>.
+		/// </summary>
+		public void TestInheritance()
+		{
+
+			FlagscriptConfigurationException fce = new FlagscriptConfigurationException();
+			Assert.IsAssignableFrom<FlagscriptException>(fce);
+
+		}
+
+	}
+
+}
